added predict yolov

+++ train_yolo.py
@@ -58,26 +58,31 @@ def train_yolo_model(output_dir, epochs):
     )
 
     # After training: load best model and evaluate
-    best_model = YOLO('runs/train/teeth-classification/weights/best.pt')
-    metrics = best_model.val(data='dataset/data.yaml')
-    # Create CSV file
-    csv_path = '/content/drive/MyDrive/yolo_output/teeth-classification/eval_metrics.csv'
+    model_path = os.path.join(output_dir,'teeth-classification/weights/best.pt')
+    if os.path.exists(model_path):
+       
+        best_model = YOLO(model_path)
+        metrics = best_model.val(data='dataset/data.yaml')
+        # Create CSV file
+        csv_path = os.path.join(output_dir,'eval_metrics.csv')
 
-    # Extract main metrics
-    rows = [
+        # Extract main metrics
+        rows = [
             ['Metric', 'Value'],
-            ['Precision', metrics.box.precision],
-            ['Recall', metrics.box.recall],
-            ['mAP50', metrics.box.map50],
-            ['mAP50-95', metrics.box.map]
-        ]
+                ['Precision', metrics.box.mp],
+                ['Recall', metrics.box.mr],
+                ['mAP50', metrics.box.map50],
+                ['mAP50-95', metrics.box.map]
+            ]
 
-        # Save to CSV
-    with open(csv_path, mode='w', newline='') as f:
-        writer = csv.writer(f)
-        writer.writerows(rows)
+            # Save to CSV
+        with open(csv_path, mode='w', newline='') as f:
+            writer = csv.writer(f)
+            writer.writerows(rows)
 
-    print(f"🔍 Metrics saved to {csv_path}")
+        print(f"🔍 Metrics saved to {csv_path}")
+    else :
+        print(f"Model {model_path} not found")
 
 
 if __name__=='__main__':
